Move native/wp pluggable wrappers to their own file
